add more information in CurrentWeather
